Combine ClientIOOffline and ClientIO classes
Moved most of the offline logic into ClientOfflineMixin instead
of using a subclass.
