+++ lib/src/client.dart
@@ -1,6 +1,6 @@
 import 'client_stub.dart'
     if (dart.library.html) 'client_browser.dart'
-    if (dart.library.io) 'client_io_offline.dart';
+    if (dart.library.io) 'client_io.dart';
 import 'enums.dart';
 import 'response.dart';
 import 'upload_progress.dart';
